# api-docs.json
{"swagger":"2.0","info":{"description":"REST API","version":"1.0.0","title":"Spring Boot REST API"},"host":"localhost:8080","basePath":"/","tags":[{"name":"Score Operations","description":"Score Controller"},{"name":"User Operations","description":"User Controller"}],"paths":{"/score/add":{"post":{"tags":["Score Operations"],"summary":"Adding new score operation.","operationId":"addScoreUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"input","description":"input","required":true,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Score"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/score/all":{"get":{"tags":["Score Operations"],"summary":"Lists all of the scores.","operationId":"getAllUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"object"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/score/leaderboard/all_time":{"get":{"tags":["Score Operations"],"summary":"Lists all of the max scores of each user.","operationId":"getBoardUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"object"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/score/leaderboard/week":{"get":{"tags":["Score Operations"],"summary":"Lists all of the weekly max scores of each user.","operationId":"getWeekBoardUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"object"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/score/max/{username}":{"get":{"tags":["Score Operations"],"summary":"Gets the max score of the requested user.","operationId":"maxScoreUsingGET","produces":["*/*"],"parameters":[{"name":"username","in":"path","description":"username","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/user/all":{"get":{"tags":["User Operations"],"summary":"Lists all of the users.","operationId":"getAllUsersUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"object"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/user/delete":{"delete":{"tags":["User Operations"],"summary":"Delete user.","operationId":"deleteUserUsingDELETE","produces":["*/*"],"parameters":[{"in":"body","name":"input","description":"input","required":true,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"User and the scores are successfully deleted.","schema":{"$ref":"#/definitions/User"}},"204":{"description":"No Content"},"401":{"description":"You are not authorized to delete the user."},"403":{"description":"Forbidden"}},"deprecated":false}},"/user/find/{username}":{"get":{"tags":["User Operations"],"summary":"Find user information corresponding to the username.","operationId":"findUserUsingGET","produces":["*/*"],"parameters":[{"name":"username","in":"path","description":"username","required":true,"type":"string"}],"responses":{"200":{"description":"Find operation is done.","schema":{"$ref":"#/definitions/User"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"User you were trying to reach is not found."}},"deprecated":false}},"/user/login":{"post":{"tags":["User Operations"],"summary":"Login operation for the users.","operationId":"loginUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"user","description":"user","required":true,"schema":{"$ref":"#/definitions/User"}}],"responses":{"200":{"description":"User is logged in successfully.","schema":{"type":"string"}},"201":{"description":"Created"},"400":{"description":"Wrong username or password."},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/user/signup":{"post":{"tags":["User Operations"],"summary":"Signup operation for users.","operationId":"addUserUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"user","description":"user","required":true,"schema":{"$ref":"#/definitions/User"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/User"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/user/update":{"put":{"tags":["User Operations"],"summary":"Update user information.","operationId":"updateUserUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"input","description":"input","required":true,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"User is successfully updated.","schema":{"$ref":"#/definitions/User"}},"201":{"description":"Created"},"401":{"description":"You are not authorized to update the user."},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}}},"definitions":{"Map«string,string»":{"type":"object","title":"Map«string,string»","additionalProperties":{"type":"string"}},"Score":{"type":"object","properties":{"date":{"type":"string","format":"date-time"},"point":{"type":"integer","format":"int32"},"score_id":{"type":"integer","format":"int64"},"user":{"$ref":"#/definitions/User"}},"title":"Score"},"User":{"type":"object","properties":{"createdDate":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"last_login":{"type":"string","format":"date-time"},"password":{"type":"string"},"salt":{"type":"string","format":"byte"},"session":{"type":"string"},"userName":{"type":"string"},"username":{"type":"string"}},"title":"User"}}}